Moved fixExtension to its own class
