+++ app/src/main/java/fr/free/nrw/commons/utils/Utils.java
@@ -7,39 +7,7 @@ import android.text.SpannableString;
 import android.text.style.UnderlineSpan;
 import android.widget.TextView;
 
-import java.util.Locale;
-import java.util.regex.Pattern;
-
 public class Utils {
-
-    /**
-     * Adds extension to filename. Converts to .jpg if system provides .jpeg, adds .jpg if no extension detected
-     * @param title File name
-     * @param extension Correct extension
-     * @return File with correct extension
-     */
-    public static String fixExtension(String title, String extension) {
-        Pattern jpegPattern = Pattern.compile("\\.jpeg$", Pattern.CASE_INSENSITIVE);
-
-        // People are used to ".jpg" more than ".jpeg" which the system gives us.
-        if (extension != null && extension.toLowerCase(Locale.ENGLISH).equals("jpeg")) {
-            extension = "jpg";
-        }
-        title = jpegPattern.matcher(title).replaceFirst(".jpg");
-        if (extension != null && !title.toLowerCase(Locale.getDefault())
-                .endsWith("." + extension.toLowerCase(Locale.ENGLISH))) {
-            title += "." + extension;
-        }
-
-        // If extension is still null, make it jpg. (Hotfix for https://github.com/commons-app/apps-android-commons/issues/228)
-        // If title has an extension in it, if won't be true
-        if (extension == null && title.lastIndexOf(".")<=0) {
-           extension = "jpg";
-           title += "." + extension;
-        }
-
-        return title;
-    }
 
     /*
     *Copies the content to the clipboard
