+++ app/src/test/kotlin/fr/free/nrw/commons/utils/UtilsFixExtensionTest.kt
@@ -1,6 +1,6 @@
 package fr.free.nrw.commons.utils
 
-import fr.free.nrw.commons.utils.Utils.fixExtension
+import fr.free.nrw.commons.utils.FilenameUtils.fixExtension
 import org.junit.Assert.assertEquals
 import org.junit.Test
